refactor: cmdとinternalディレクトリに分離
