+++ typing-server/internal/handler/health.go
