+++ typing-server/internal/middleware/cors.go
