+++ typing-server/internal/middleware/trace.go
