+++ typing-server/internal/service/user.go
@@ -3,7 +3,7 @@ package service
 import (
 	"context"
 
-	"github.com/su-its/typing/typing-server/api/repository"
+	"github.com/su-its/typing/typing-server/internal/repository"
 	"github.com/su-its/typing/typing-server/domain/repository/ent"
 )
